+++ features/plebhy/Views/index.js
@@ -1,0 +1,2 @@
+export {default as PlebhyGifView} from './PlebhyGifView';
+export {default as PlebhyStickerView} from './PlebhyStickerView';
