+++ features/plebhy/components/index.js
@@ -1,0 +1,1 @@
+export {default as GifContainer} from './GifContainer';
