+++ features/plebhy/index.js
@@ -1,0 +1,1 @@
+export {default as PlebhyNavigator} from './nav/PlebhyNavigator'
